breaking up larger component to more consise
